requirements.txt
flask==1.0.2
gunicorn==19.9.0
h5py<3.0.0
tensorflow==1.12.0
Pillow
numpy
